+++ python/tests/test_problem.py
@@ -225,7 +225,9 @@ class TestProblem(unittest.TestCase):
             coeff = np.random.randn(f.shape[0])
             p.eval(x0)
             self.assertRaises(pf.ProblemError,p.combine_H,np.zeros(f.shape[0]+1),False)
+            self.assertTrue(p.has_error())
             p.clear_error()
+            self.assertFalse(p.has_error())
             p.combine_H(coeff,False)
             J0 = p.J.copy()
             g0 = J0.T*coeff
@@ -508,7 +510,9 @@ class TestProblem(unittest.TestCase):
             coeff = np.random.randn(f.shape[0])
             p.eval(x0)
             self.assertRaises(pf.ProblemError,p.combine_H,np.zeros(f.shape[0]+1),False)
+            self.assertTrue(p.has_error())
             p.clear_error()
+            self.assertFalse(p.has_error())
             p.combine_H(coeff,False)
             J0 = p.J.copy()
             g0 = J0.T*coeff